+++ examples/examples/starwars.rs
@@ -23,7 +23,7 @@ struct FilmArguments {
 #[cynic(
     schema_path = "../schemas/starwars.schema.graphql",
     graphql_type = "Root",
-    argument_struct = "FilmArguments"
+    variables = "FilmArguments"
 )]
 struct FilmDirectorQuery {
     #[arguments(id: $id)]